requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.0.3
django-debug-toolbar==1.9.1
django-filter==2.0.0
django-mail-templated==2.6.5
django-modeladmin-reorder==0.3.1
django-uuslug==1.1.8
djangorestframework==3.7.7
djangorestframework-filters==0.10.2
facebook-sdk==3.0.0
google-api-python-client==1.6.5
httplib2==0.11.0
idna==2.7
mysqlclient==1.3.12
oauth2client==4.1.2
Pillow==5.0.0
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-slugify==1.2.5
pytz==2018.3
requests==2.19.1
rsa==3.4.2
six==1.11.0
sqlparse==0.2.4
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.23
uWSGI==2.0.17
Wand==0.4.4
